backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
httpx==0.25.2
google-generativeai==0.3.0
pymongo==4.6.0
motor==3.3.2
python-multipart==0.0.6
reportlab==4.0.7
python-csv==0.0.13

